device_manager/requirements.txt
Django
djangorestframework
drf_yasg
django-filter
uwsgi